lucia_backend/pyproject.toml
[project]
name = "lucia-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "sungminna",email = "sungmin.na330@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "delos-lucia (>=0.1.1,<0.2.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "daphne (>=4.1.2,<5.0.0)",
    "channels (>=4.2.2,<5.0.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
